Add a bridge ResponseManager for the middleware
