+++ src/Manager/ResponseManager.php
@@ -5,16 +5,13 @@ declare(strict_types=1);
 namespace Chubbyphp\ApiHttp\Manager;
 
 use Chubbyphp\ApiHttp\ApiProblem\ApiProblemInterface;
-use Chubbyphp\ApiHttp\ApiProblem\ClientError\NotAcceptable;
-use Chubbyphp\ApiHttp\ApiProblem\ClientError\UnsupportedMediaType;
-use Chubbyphp\ApiHttp\Middleware\AcceptAndContentTypeMiddlewareResponseFactoryInterface;
 use Chubbyphp\Deserialization\DeserializerInterface;
 use Chubbyphp\Serialization\Normalizer\NormalizerContextInterface;
 use Chubbyphp\Serialization\SerializerInterface;
 use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\ResponseInterface;
 
-final class ResponseManager implements ResponseManagerInterface, AcceptAndContentTypeMiddlewareResponseFactoryInterface
+final class ResponseManager implements ResponseManagerInterface
 {
     /**
      * @var ResponseFactoryInterface
@@ -104,21 +101,5 @@ final class ResponseManager implements ResponseManagerInterface, AcceptAndConten
     private function setSerializer(SerializerInterface $serializer): void
     {
         $this->serializer = $serializer;
-    }
-
-    public function createForNotAcceptable(
-        string $accept,
-        array $acceptableMimeTypes,
-        string $mimeType
-    ): ResponseInterface {
-        return $this->createFromApiProblem(new NotAcceptable($accept, $acceptableMimeTypes), $mimeType);
-    }
-
-    public function createForUnsupportedMediaType(
-        string $mediaType,
-        array $supportedMediaTypes,
-        string $mimeType
-    ): ResponseInterface {
-        return $this->createFromApiProblem(new UnsupportedMediaType($mediaType, $supportedMediaTypes), $mimeType);
     }
 }
